BREAKING CHANGE: change `work` method signature to `async computeWork` so that there's no need to call `init` anymore
