+++ docs/enums/unit.html
@@ -969,6 +969,9 @@ img { max-width: 100%; }
 			<a href="../index.html#checkwork" class="tsd-kind-icon">check<wbr>Work</a>
 		</li>
 		<li class=" tsd-kind-function tsd-is-external">
+			<a href="../index.html#computework" class="tsd-kind-icon">compute<wbr>Work</a>
+		</li>
+		<li class=" tsd-kind-function tsd-is-external">
 			<a href="../index.html#convert" class="tsd-kind-icon">convert</a>
 		</li>
 		<li class=" tsd-kind-function tsd-is-external">
@@ -990,12 +993,6 @@ img { max-width: 100%; }
 			<a href="../index.html#hashblock" class="tsd-kind-icon">hash<wbr>Block</a>
 		</li>
 		<li class=" tsd-kind-function tsd-is-external">
-			<a href="../index.html#init" class="tsd-kind-icon">init</a>
-		</li>
-		<li class=" tsd-kind-function tsd-is-external">
-			<a href="../index.html#isready" class="tsd-kind-icon">is<wbr>Ready</a>
-		</li>
-		<li class=" tsd-kind-function tsd-is-external">
 			<a href="../index.html#signblock" class="tsd-kind-icon">sign<wbr>Block</a>
 		</li>
 		<li class=" tsd-kind-function tsd-is-external">
@@ -1003,9 +1000,6 @@ img { max-width: 100%; }
 		</li>
 		<li class=" tsd-kind-function tsd-is-external">
 			<a href="../index.html#verifyblock" class="tsd-kind-icon">verify<wbr>Block</a>
-		</li>
-		<li class=" tsd-kind-function tsd-is-external">
-			<a href="../index.html#work" class="tsd-kind-icon">work</a>
 		</li>
 	</ul>
 </nav>
@@ -1046,7 +1040,7 @@ img { max-width: 100%; }
 				<div class="tsd-signature tsd-kind-icon">KNano<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol"> =&nbsp;&quot;KNano&quot;</span></div>
 				<aside class="tsd-sources">
 					<ul>
-						<li>Defined in <a href="https://github.com/marvinroger/nanocurrency-js/blob/5dc3d3c/src/conversion.ts#L39">conversion.ts:39</a></li>
+						<li>Defined in <a href="https://github.com/marvinroger/nanocurrency-js/blob/514c110/src/conversion.ts#L39">conversion.ts:39</a></li>
 					</ul>
 				</aside>
 				<div class="tsd-comment tsd-typography">
@@ -1061,7 +1055,7 @@ img { max-width: 100%; }
 				<div class="tsd-signature tsd-kind-icon">MNano<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol"> =&nbsp;&quot;MNano&quot;</span></div>
 				<aside class="tsd-sources">
 					<ul>
-						<li>Defined in <a href="https://github.com/marvinroger/nanocurrency-js/blob/5dc3d3c/src/conversion.ts#L41">conversion.ts:41</a></li>
+						<li>Defined in <a href="https://github.com/marvinroger/nanocurrency-js/blob/514c110/src/conversion.ts#L41">conversion.ts:41</a></li>
 					</ul>
 				</aside>
 				<div class="tsd-comment tsd-typography">
@@ -1076,7 +1070,7 @@ img { max-width: 100%; }
 				<div class="tsd-signature tsd-kind-icon">NANO<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol"> =&nbsp;&quot;NANO&quot;</span></div>
 				<aside class="tsd-sources">
 					<ul>
-						<li>Defined in <a href="https://github.com/marvinroger/nanocurrency-js/blob/5dc3d3c/src/conversion.ts#L37">conversion.ts:37</a></li>
+						<li>Defined in <a href="https://github.com/marvinroger/nanocurrency-js/blob/514c110/src/conversion.ts#L37">conversion.ts:37</a></li>
 					</ul>
 				</aside>
 				<div class="tsd-comment tsd-typography">
@@ -1091,7 +1085,7 @@ img { max-width: 100%; }
 				<div class="tsd-signature tsd-kind-icon">Nano<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol"> =&nbsp;&quot;Nano&quot;</span></div>
 				<aside class="tsd-sources">
 					<ul>
-						<li>Defined in <a href="https://github.com/marvinroger/nanocurrency-js/blob/5dc3d3c/src/conversion.ts#L35">conversion.ts:35</a></li>
+						<li>Defined in <a href="https://github.com/marvinroger/nanocurrency-js/blob/514c110/src/conversion.ts#L35">conversion.ts:35</a></li>
 					</ul>
 				</aside>
 				<div class="tsd-comment tsd-typography">
@@ -1106,7 +1100,7 @@ img { max-width: 100%; }
 				<div class="tsd-signature tsd-kind-icon">hex<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol"> =&nbsp;&quot;hex&quot;</span></div>
 				<aside class="tsd-sources">
 					<ul>
-						<li>Defined in <a href="https://github.com/marvinroger/nanocurrency-js/blob/5dc3d3c/src/conversion.ts#L27">conversion.ts:27</a></li>
+						<li>Defined in <a href="https://github.com/marvinroger/nanocurrency-js/blob/514c110/src/conversion.ts#L27">conversion.ts:27</a></li>
 					</ul>
 				</aside>
 				<div class="tsd-comment tsd-typography">
@@ -1121,7 +1115,7 @@ img { max-width: 100%; }
 				<div class="tsd-signature tsd-kind-icon">knano<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol"> =&nbsp;&quot;knano&quot;</span></div>
 				<aside class="tsd-sources">
 					<ul>
-						<li>Defined in <a href="https://github.com/marvinroger/nanocurrency-js/blob/5dc3d3c/src/conversion.ts#L33">conversion.ts:33</a></li>
+						<li>Defined in <a href="https://github.com/marvinroger/nanocurrency-js/blob/514c110/src/conversion.ts#L33">conversion.ts:33</a></li>
 					</ul>
 				</aside>
 				<div class="tsd-comment tsd-typography">
@@ -1136,7 +1130,7 @@ img { max-width: 100%; }
 				<div class="tsd-signature tsd-kind-icon">nano<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol"> =&nbsp;&quot;nano&quot;</span></div>
 				<aside class="tsd-sources">
 					<ul>
-						<li>Defined in <a href="https://github.com/marvinroger/nanocurrency-js/blob/5dc3d3c/src/conversion.ts#L31">conversion.ts:31</a></li>
+						<li>Defined in <a href="https://github.com/marvinroger/nanocurrency-js/blob/514c110/src/conversion.ts#L31">conversion.ts:31</a></li>
 					</ul>
 				</aside>
 				<div class="tsd-comment tsd-typography">
@@ -1151,7 +1145,7 @@ img { max-width: 100%; }
 				<div class="tsd-signature tsd-kind-icon">raw<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol"> =&nbsp;&quot;raw&quot;</span></div>
 				<aside class="tsd-sources">
 					<ul>
-						<li>Defined in <a href="https://github.com/marvinroger/nanocurrency-js/blob/5dc3d3c/src/conversion.ts#L29">conversion.ts:29</a></li>
+						<li>Defined in <a href="https://github.com/marvinroger/nanocurrency-js/blob/514c110/src/conversion.ts#L29">conversion.ts:29</a></li>
 					</ul>
 				</aside>
 				<div class="tsd-comment tsd-typography">
