+++ docs/interfaces/blockdata.html
@@ -960,6 +960,9 @@ img { max-width: 100%; }
 			<a href="../index.html#checkwork" class="tsd-kind-icon">check<wbr>Work</a>
 		</li>
 		<li class=" tsd-kind-function tsd-is-external">
+			<a href="../index.html#computework" class="tsd-kind-icon">compute<wbr>Work</a>
+		</li>
+		<li class=" tsd-kind-function tsd-is-external">
 			<a href="../index.html#convert" class="tsd-kind-icon">convert</a>
 		</li>
 		<li class=" tsd-kind-function tsd-is-external">
@@ -981,12 +984,6 @@ img { max-width: 100%; }
 			<a href="../index.html#hashblock" class="tsd-kind-icon">hash<wbr>Block</a>
 		</li>
 		<li class=" tsd-kind-function tsd-is-external">
-			<a href="../index.html#init" class="tsd-kind-icon">init</a>
-		</li>
-		<li class=" tsd-kind-function tsd-is-external">
-			<a href="../index.html#isready" class="tsd-kind-icon">is<wbr>Ready</a>
-		</li>
-		<li class=" tsd-kind-function tsd-is-external">
 			<a href="../index.html#signblock" class="tsd-kind-icon">sign<wbr>Block</a>
 		</li>
 		<li class=" tsd-kind-function tsd-is-external">
@@ -994,9 +991,6 @@ img { max-width: 100%; }
 		</li>
 		<li class=" tsd-kind-function tsd-is-external">
 			<a href="../index.html#verifyblock" class="tsd-kind-icon">verify<wbr>Block</a>
-		</li>
-		<li class=" tsd-kind-function tsd-is-external">
-			<a href="../index.html#work" class="tsd-kind-icon">work</a>
 		</li>
 	</ul>
 </nav>
@@ -1042,7 +1036,7 @@ img { max-width: 100%; }
 				<div class="tsd-signature tsd-kind-icon">balance<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 				<aside class="tsd-sources">
 					<ul>
-						<li>Defined in <a href="https://github.com/marvinroger/nanocurrency-js/blob/5dc3d3c/src/block.ts#L24">block.ts:24</a></li>
+						<li>Defined in <a href="https://github.com/marvinroger/nanocurrency-js/blob/514c110/src/block.ts#L24">block.ts:24</a></li>
 					</ul>
 				</aside>
 				<div class="tsd-comment tsd-typography">
@@ -1057,7 +1051,7 @@ img { max-width: 100%; }
 				<div class="tsd-signature tsd-kind-icon">link<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 				<aside class="tsd-sources">
 					<ul>
-						<li>Defined in <a href="https://github.com/marvinroger/nanocurrency-js/blob/5dc3d3c/src/block.ts#L26">block.ts:26</a></li>
+						<li>Defined in <a href="https://github.com/marvinroger/nanocurrency-js/blob/514c110/src/block.ts#L26">block.ts:26</a></li>
 					</ul>
 				</aside>
 				<div class="tsd-comment tsd-typography">
@@ -1072,7 +1066,7 @@ img { max-width: 100%; }
 				<div class="tsd-signature tsd-kind-icon">previous<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 				<aside class="tsd-sources">
 					<ul>
-						<li>Defined in <a href="https://github.com/marvinroger/nanocurrency-js/blob/5dc3d3c/src/block.ts#L20">block.ts:20</a></li>
+						<li>Defined in <a href="https://github.com/marvinroger/nanocurrency-js/blob/514c110/src/block.ts#L20">block.ts:20</a></li>
 					</ul>
 				</aside>
 				<div class="tsd-comment tsd-typography">
@@ -1087,7 +1081,7 @@ img { max-width: 100%; }
 				<div class="tsd-signature tsd-kind-icon">representative<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 				<aside class="tsd-sources">
 					<ul>
-						<li>Defined in <a href="https://github.com/marvinroger/nanocurrency-js/blob/5dc3d3c/src/block.ts#L22">block.ts:22</a></li>
+						<li>Defined in <a href="https://github.com/marvinroger/nanocurrency-js/blob/514c110/src/block.ts#L22">block.ts:22</a></li>
 					</ul>
 				</aside>
 				<div class="tsd-comment tsd-typography">
@@ -1102,7 +1096,7 @@ img { max-width: 100%; }
 				<div class="tsd-signature tsd-kind-icon">work<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">null</span></div>
 				<aside class="tsd-sources">
 					<ul>
-						<li>Defined in <a href="https://github.com/marvinroger/nanocurrency-js/blob/5dc3d3c/src/block.ts#L18">block.ts:18</a></li>
+						<li>Defined in <a href="https://github.com/marvinroger/nanocurrency-js/blob/514c110/src/block.ts#L18">block.ts:18</a></li>
 					</ul>
 				</aside>
 				<div class="tsd-comment tsd-typography">
