+++ docs/interfaces/convertparams.html
@@ -951,6 +951,9 @@ img { max-width: 100%; }
 			<a href="../index.html#checkwork" class="tsd-kind-icon">check<wbr>Work</a>
 		</li>
 		<li class=" tsd-kind-function tsd-is-external">
+			<a href="../index.html#computework" class="tsd-kind-icon">compute<wbr>Work</a>
+		</li>
+		<li class=" tsd-kind-function tsd-is-external">
 			<a href="../index.html#convert" class="tsd-kind-icon">convert</a>
 		</li>
 		<li class=" tsd-kind-function tsd-is-external">
@@ -972,12 +975,6 @@ img { max-width: 100%; }
 			<a href="../index.html#hashblock" class="tsd-kind-icon">hash<wbr>Block</a>
 		</li>
 		<li class=" tsd-kind-function tsd-is-external">
-			<a href="../index.html#init" class="tsd-kind-icon">init</a>
-		</li>
-		<li class=" tsd-kind-function tsd-is-external">
-			<a href="../index.html#isready" class="tsd-kind-icon">is<wbr>Ready</a>
-		</li>
-		<li class=" tsd-kind-function tsd-is-external">
 			<a href="../index.html#signblock" class="tsd-kind-icon">sign<wbr>Block</a>
 		</li>
 		<li class=" tsd-kind-function tsd-is-external">
@@ -985,9 +982,6 @@ img { max-width: 100%; }
 		</li>
 		<li class=" tsd-kind-function tsd-is-external">
 			<a href="../index.html#verifyblock" class="tsd-kind-icon">verify<wbr>Block</a>
-		</li>
-		<li class=" tsd-kind-function tsd-is-external">
-			<a href="../index.html#work" class="tsd-kind-icon">work</a>
 		</li>
 	</ul>
 </nav>
@@ -1030,7 +1024,7 @@ img { max-width: 100%; }
 				<div class="tsd-signature tsd-kind-icon">from<span class="tsd-signature-symbol">:</span> <a href="../enums/unit.html" class="tsd-signature-type">Unit</a></div>
 				<aside class="tsd-sources">
 					<ul>
-						<li>Defined in <a href="https://github.com/marvinroger/nanocurrency-js/blob/5dc3d3c/src/conversion.ts#L47">conversion.ts:47</a></li>
+						<li>Defined in <a href="https://github.com/marvinroger/nanocurrency-js/blob/514c110/src/conversion.ts#L47">conversion.ts:47</a></li>
 					</ul>
 				</aside>
 				<div class="tsd-comment tsd-typography">
@@ -1045,7 +1039,7 @@ img { max-width: 100%; }
 				<div class="tsd-signature tsd-kind-icon">to<span class="tsd-signature-symbol">:</span> <a href="../enums/unit.html" class="tsd-signature-type">Unit</a></div>
 				<aside class="tsd-sources">
 					<ul>
-						<li>Defined in <a href="https://github.com/marvinroger/nanocurrency-js/blob/5dc3d3c/src/conversion.ts#L49">conversion.ts:49</a></li>
+						<li>Defined in <a href="https://github.com/marvinroger/nanocurrency-js/blob/514c110/src/conversion.ts#L49">conversion.ts:49</a></li>
 					</ul>
 				</aside>
 				<div class="tsd-comment tsd-typography">
