+++ docs/interfaces/deriveaddressparams.html
@@ -948,6 +948,9 @@ img { max-width: 100%; }
 			<a href="../index.html#checkwork" class="tsd-kind-icon">check<wbr>Work</a>
 		</li>
 		<li class=" tsd-kind-function tsd-is-external">
+			<a href="../index.html#computework" class="tsd-kind-icon">compute<wbr>Work</a>
+		</li>
+		<li class=" tsd-kind-function tsd-is-external">
 			<a href="../index.html#convert" class="tsd-kind-icon">convert</a>
 		</li>
 		<li class=" tsd-kind-function tsd-is-external">
@@ -969,12 +972,6 @@ img { max-width: 100%; }
 			<a href="../index.html#hashblock" class="tsd-kind-icon">hash<wbr>Block</a>
 		</li>
 		<li class=" tsd-kind-function tsd-is-external">
-			<a href="../index.html#init" class="tsd-kind-icon">init</a>
-		</li>
-		<li class=" tsd-kind-function tsd-is-external">
-			<a href="../index.html#isready" class="tsd-kind-icon">is<wbr>Ready</a>
-		</li>
-		<li class=" tsd-kind-function tsd-is-external">
 			<a href="../index.html#signblock" class="tsd-kind-icon">sign<wbr>Block</a>
 		</li>
 		<li class=" tsd-kind-function tsd-is-external">
@@ -982,9 +979,6 @@ img { max-width: 100%; }
 		</li>
 		<li class=" tsd-kind-function tsd-is-external">
 			<a href="../index.html#verifyblock" class="tsd-kind-icon">verify<wbr>Block</a>
-		</li>
-		<li class=" tsd-kind-function tsd-is-external">
-			<a href="../index.html#work" class="tsd-kind-icon">work</a>
 		</li>
 	</ul>
 </nav>
@@ -1026,7 +1020,7 @@ img { max-width: 100%; }
 				<div class="tsd-signature tsd-kind-icon">use<wbr>Nano<wbr>Prefix<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">undefined</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">true</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">false</span></div>
 				<aside class="tsd-sources">
 					<ul>
-						<li>Defined in <a href="https://github.com/marvinroger/nanocurrency-js/blob/5dc3d3c/src/keys.ts#L91">keys.ts:91</a></li>
+						<li>Defined in <a href="https://github.com/marvinroger/nanocurrency-js/blob/514c110/src/keys.ts#L88">keys.ts:88</a></li>
 					</ul>
 				</aside>
 				<div class="tsd-comment tsd-typography">
